+++ main/CMakeLists.txt
@@ -18,6 +18,7 @@ SRCS
     "./tasks/user_input_task.c"
     "./tasks/power_management_task.c"
     "./http_server/http_server.c"
+    "./self_test/self_test.c"
 
 INCLUDE_DIRS
     "."
