+++ main/self_test/self_text.h
@@ -1,0 +1,6 @@
+#ifndef SELF_TEST_H_
+#define SELF_TEST_H_
+
+void self_test(void * pvParameters);
+
+#endif
